requirements.txt
fusepy>=2.0.0